feat: replace old provider impls with Genkit adapters (Task 6)
Delete 26 hand-rolled provider files and replace their usage in
provider_registry.go, module_provider.go, and step_model_pull.go
with Genkit factory calls. Preserve anthropic_bedrock.go (still used
by genkit/providers.go) alongside new anthropic_bedrock_convert.go.
Add OllamaClient for Pull/ListModels utility access in step_model_pull.
Move models.go constants/types/helpers from deleted files.
